BOAC-6176, /degrees page: one table for non-archived and one for archived
